Update Lichess_Topbar_Menu.user.js to version 1.3
Version 1.3 
Added background color menu for light/dark theme.
Added the new hierarchy menu and the previous one as optional. (Switch the setting true/false)
Forum part: Uses the full width of the site when you are on the forum which involves hiding the useless and empty sidebar and the logo.
Tournament part: Added feature to always  be rated selected when make new tournament.